INITIAL COMMIT: Introduce Cargo, move all the `bitcoin-rs` stuff to its own repo
The Bitcoin library is now available at
  https://github.com/apoelstra/bitcoin-rs

This repository now only covers the Wizards' Wallet. To build it,
obtain Cargo from crates.io then run

  cargo build
  cargo test

In the root directory of this project.

Oh, and I'm declaring this the unofficial 'Initial commit'. Had I not
published this code so prematurely (damn Github being a free backup),
I would squash everything into this commit. But I'll not, to avoid
doing a `push -f`.